Split site add utilities to Setup directory
